requirements.txt
altgraph==0.17.5
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pefile==2024.8.26
psycopg2-binary==2.9.11
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0